app/requirements.txt
flask
flask_login
python-dotenv
requests
oauthlib
pyOpenSSL
